storiahackathon/requirement.txt
MySQL-python==1.2.4
py-bcrypt==0.4
simplejson==3.3.0
tornado==3.1.1
torndb==0.1
wsgiref==0.1.2
